issue #75 refactor: axios instance name change
