+++ src/api/auth/postEmailAuthSend.ts
@@ -1,6 +1,6 @@
 import { AxiosResponse } from "axios";
 
-import { axiosInstance } from "@/api/customAxios";
+import { axiosInstance } from "@/api/axiosInstance";
 
 import { END_POINTS } from "@/constants/api";
