+++ src/api/auth/postSignUp.ts
@@ -1,6 +1,6 @@
 import type { AxiosResponse } from "axios";
 
-import { axiosInstance } from "@/api/customAxios";
+import { axiosInstance } from "@/api/axiosInstance";
 
 import { END_POINTS } from "@/constants/api";
