+++ src/hooks/api/useLogInMutation.ts
@@ -8,7 +8,7 @@ import { isLoggedInState } from "@/store/auth";
 import { TokenType } from "@/types/auth";
 
 import { postLogIn } from "@/api/auth/postLogin";
-import { authorizedAxiosInstance } from "@/api/customAxios";
+import { authorizedAxiosInstance } from "@/api/axiosInstance";
 
 export const useLogInMutation = () => {
 	const setIsLoggedIn = useSetRecoilState(isLoggedInState);
